requirements.txt
pillow
requests
numpy
scipy
matplotlib
scikit-learn
tensorflow
torch